Casts to avoid integer overflow in matrix row/columns. (#1281)
* Casts to avoid integer overflow in matrix row/columns.

* Updated ChangeLog.

* Missed a few spots.

* Missed another spot.

* Even more spots missed.

* Remove an extraneous cast that breaks the Vector::end iterator.

* Updated Changelog again.

+++ inst/include/Rcpp/vector/Vector.h
@@ -331,7 +331,7 @@ public:
     }
 
     inline iterator begin() { return cache.get() ; }
-    inline iterator end() { return cache.get() + static_cast<int>(size()) ; }
+    inline iterator end() { return cache.get() + size() ; }
     inline const_iterator begin() const{ return cache.get_const() ; }
     inline const_iterator end() const{ return cache.get_const() + size() ; }
     inline const_iterator cbegin() const{ return cache.get_const() ; }
